+++ CHANGELOG.md
@@ -8,16 +8,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - SFTP support
 - `.stat()` method that return information abouth the given path
 - Connector constructor now accepts `infos` to override static infos
+- FS connector now has a sandbox to restrict file access
 
 ### Fixed
 - `.getInfos()` now correctly get the session
 - Takes GitHub `redirectUri` into account for OAuth requests
-- Fix Dropbox batch operation
+- Fixes Dropbox batch operation
+- Fixes Dropbox infos when account has not been set yet
 
 ### Changed
 - Dropbox is now more tolerant with the path (first `/` is optional)
-- Include connector documentation to the global one
-- Include README in documentation
+- Includes connector documentation to the global one
+- Includes README in documentation
 
 ## [1.1.0] - 2017-04-26
 ### Added